# Make `gix_path::env:shell()` more robust and use it in `gix-command`