--- a/gix-path/src/env/auxiliary.rs
+++ b/gix-path/src/env/auxiliary.rs
@@ -71,10 +71,10 @@ const BIN_DIR_FRAGMENTS: &[&str] = &["bin", "usr/bin"];
 /// does, which is the case unless it is overridden by setting `GIT_EXEC_PATH` to an unusual value.
 ///
 /// This is currently only used (and only exercised in tests) for finding `sh.exe`. It may be used
-/// to find other executables in the future, but may require adjustment. (In particular, depending
-/// on the desired semantics, it should possibly also check inside a `cmd` directory, possibly also
-/// check inside `super::core_dir()` itself, and could safely check the latter location even if its
-/// value is not suitable to use in inferring any other paths.)
+/// to find other executables in the future, but may require adjustment. In particular, depending
+/// on the desired semantics, it should possibly also check inside a `cmd` directory; directories
+/// like `<platform>/bin`, for any applicable variants (such as `mingw64`); and `super::core_dir()`
+/// itself, which it could safely check even if its value is not safe for inferring other paths.
 fn find_git_associated_windows_executable(stem: &str) -> Option<OsString> {
     let git_root = git_for_windows_root()?;