--- a/gix-path/src/env/auxiliary.rs
+++ b/gix-path/src/env/auxiliary.rs
@@ -45,13 +45,25 @@ const MSYS_USR_VARIANTS: &[&str] = &["mingw64", "mingw32", "clangarm64", "clang6
 ///
 /// Currently this is used only for finding the path to an `sh.exe` associated with Git. This is
 /// separate from `installation_config()` and `installation_config_prefix()` in `gix_path::env`.
+/// This is *not* suitable for finding the highest-scoped configuration file, because that could be
+/// installed in an unusual place, or customized via `GIT_CONFIG_SYSTEM` or `GIT_CONFIG_NOSYSTEM`,
+/// all of which `installation_config()` should reflect. Likewise, `installation_config_prefix()`
+/// has strong uses, such as to find a directory inside `ProgramData` containing configuration.
+/// But it is possible that some marginal uses of `installation_config_prefix()`, if they do not
+/// really relate to configuration, could be replaced with `git_for_windows_root()` in the future.
 fn git_for_windows_root() -> Option<&'static Path> {
     static GIT_ROOT: Lazy<Option<PathBuf>> = Lazy::new(|| {
         super::core_dir()
-            .filter(|core| core.is_absolute() && core.ends_with("libexec/git-core"))
+            .filter(|core| {
+                // Only use this if the directory structure resembles a Git installation. This
+                // accepts installations of common types that are not broken when `GIT_EXEC_PATH`
+                // is unset, as well as values of `GIT_EXEC_PATH` that are likely to be usable.
+                core.is_absolute() && core.ends_with("libexec/git-core")
+            })
             .and_then(|core| core.ancestors().nth(2))
             .filter(|prefix| {
                 // Only use `libexec/git-core` from inside something `usr`-like, such as `mingw64`.
+                // See `MSYS_USR_VARIANTS` for details and the rationale for this restriction.
                 MSYS_USR_VARIANTS.iter().any(|name| prefix.ends_with(name))
             })
             .and_then(|prefix| prefix.parent())