… Cargo.lock
… gix-command/tests/command.rs
… gix-credentials/tests/program/from_custom_definition.rs