Tasks since https://github.com/GitoxideLabs/gitoxide/pull/1862#issuecomment-2692158831:

- [x] Use the shim `sh.exe` when available
- [x] Fall back to the non-shim `sh.exe`
- [x] Test locally to ensure the failures still go away with the shim
- [x] Only use inferred `(git root)` if it looks like Git for Windows and looks safe
- [x] Fix unrelated `test-fixtures-windows` breakage ([#1849](https://github.com/GitoxideLabs/gitoxide/issues/1849), [#1870](https://github.com/GitoxideLabs/gitoxide/pull/1870)) and rebase
- [x] Fix unrelated `test-32bit` breakage ([#1874](https://github.com/GitoxideLabs/gitoxide/pull/1874)) and rebase
- [x] Retest locally after all nontrivial rebasing
- [x] Investigate [local failures](https://github.com/GitoxideLabs/gitoxide/pull/1864#issuecomment-2724464883) in [#1864](https://github.com/GitoxideLabs/gitoxide/pull/1864) with non-shim, in case relevant here

Outscoped:

- Report the observed `>>` problem as a bug to MSYS2 or a related project
- https://github.com/GitoxideLabs/gitoxide/issues/1868

(On all the above, see "What it means for this PR" in https://github.com/GitoxideLabs/gitoxide/pull/1862#issuecomment-2692158831.)

The original PR description follows.

---

This makes `git_path::env::shell()` more likely to work correctly on Windows, then attempts to use it in `gix_command::Prepare` when running a command with a shell due to `use_shell` being set to `true`. Full details are in the commit messages, though the changes are also summarized here.

I think this is not ready to merge yet, due to failures that happen when running the tests locally from PowerShell, but that I cannot produce otherwise. I do not know what causes these failures, and I expect I may have difficulty getting this to a point where it would be ready to merge without input. Since the failures happen in a test that was fairly recently added, I figured you might just know what is going on.

I think `shell()` can and should be used for `gix-command`. But if not then the reason will be relevant to other potential uses of `shell()`. In that case, assuming they leave enough good uses for `shell()` that it is kept, I hope to discover and articulate those reasons in the documentation for `shell()`. But I think most likely there is a bug or wrong test expectation somewhere that accounts for the local test failure, and that either the current implementation of `shell()` with the modifications included here, or something much like it, is suitable for broad use.

#### Making `gix_path::env::shell()` more robust

This makes `gix_path::env::shell()` more likely to work correctly on Windows by:

- Checking that the path to a `sh.exe` associated with Git for Windows actually exists (and is either a regular file or a symbolic link that, when fully dereferenced, gives a file), because if that is not the case, then the fallback of just looking up `sh.exe` using a `PATH` search is more likely to succeed.

  This change is needed to use `shell()` more widely without breaking things on systems where `sh.exe` is available and usable but cannot be found in the usual location where Git for Windows supplies it.

- Using `/` rather than `\` directory separators when appending `usr`, `bin`, and `sh.exe` components to the Git for Windows installation directory path obtained via `git --exec-path` and going up three components. This causes the path through `sh.exe` is run to use all `/` separators except in the unusual situation that `GIT_EXEC_PATH` has been set to a path with `\` separator. (This is unusual because people don't usually set `GIT_EXEC_PATH`. It would probably be best, when setting it on Windows, to use `/` separators, but I do not actually know how often people do that.)

  I believe this is only a minor improvement, in view of this path not being automatically passed through to the shell, as described in [#1840 (reply in thread)](https://github.com/GitoxideLabs/gitoxide/discussions/1840#discussioncomment-12294328). But this change also makes tests in `gix-command` and `gix-credentials` pass that would otherwise require more extensive modification related to `\` escaping, when modifying `gix-command` to use `shell()`.

#### Using `gix_path::env::shell()` in `gix-command`

Before the changes in this PR, `gix-command` uses the relative path `sh.exe` on Windows. This works a significant fraction of the time already. In particular, the problems with the relative path `bash.exe` that make it frequently unusable to find a non-WSL `bash` on Windows, described in https://github.com/GitoxideLabs/gitoxide/issues/1359#issuecomment-2316614616, do not apply to `sh.exe`, since there is no Microsoft-provided `sh.exe` related to WSL. However:

- Such an `sh.exe` is not always the best choice when present.
- There may be no `sh.exe` that can be found in `PATH`. For example, if the only `sh.exe` is supplied by Git for Windows, and neither of the Git for Windows directories that contain an `sh.exe` binary are in `PATH`, then before the changes in this PR, a shell will not be found unless `shell_program()` is called (every time a command with `use_shell` is run).

Therefore, this uses `shell()`, with the improvements described above, in `gix-command`.

---

As alluded to above, I think this is not ready yet. All tests pass on CI, as well a when run locally from Git Bash. But when running them locally in PowerShell, `gix-merge::merge blob::platform::merge::with_external` fails. The failure does not require `GIX_TEST_IGNORE_ARCHIVES` to be set, and this PR does not currently modify any test fixture scripts.

Windows failures that occur when tests are run from PowerShell but not from Git Bash have usually been due to #1359 and are thus not expected since #1712, where `gix-testtools` finds `bash.exe` associated with Git for Windows and uses it to run fixtures. But that fix didn't affect the way `gix-command` uses a shell, so it does not provide for the kind of shell invocations occurring here.

Still, it doesn't make sense to me why the failure happens, and why it happens only when I run the tests locally from PowerShell. On this system, in the `PATH`, including when accessed from PowerShell, the command `sh` finds is a shim for the same shell that, as of this PR, `gix-path` finds and `gix-command` uses. Furthermore, the change here makes `gix-path` find `sh.exe` *more like* the way `gix-testtools` finds `bash.exe`, both of which are Bash when provided by Git for Windows.

Furthermore, while most of my local testing has been with Git for Windows 2.48.1, I have verified that the test fails the same way on the same system with Git for Windows 2.47.1(2), and that the experiment described below to examine the details of the failure also gives the same results with that version. This is to say that it seems like this failure is *not* related, even conceptually, to #1849.

```text
C:\Users\ek\source\repos\gitoxide [run-ci/consistent-sh ≡]> cargo nextest run -p gix-merge --no-fail-fast
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.48s
────────────
 Nextest run ID 5771b786-7fff-43d7-9721-b20970cbb558 with nextest profile: default
    Starting 20 tests across 2 binaries
     Running [ 00:00:00] 0/20: 0 running, 0 passed, 0 skipped
        PASS [   0.018s] gix-merge::merge blob::builtin_driver::text::fuzzed
        PASS [   0.022s] gix-merge::merge blob::builtin_driver::text::both_sides_same_changes_are_conflict_free
        PASS [   0.025s] gix-merge::merge blob::pipeline::without_transformation
        PASS [   0.031s] gix-merge::merge blob::pipeline::non_existing
        PASS [   0.035s] gix-merge::merge blob::pipeline::binary_below_large_file_threshold
        PASS [   0.038s] gix-merge::merge blob::builtin_driver::text::both_differ_partially_resolution_is_conflicting
        PASS [   0.040s] gix-merge::merge blob::builtin_driver::binary
        PASS [   0.047s] gix-merge::merge blob::pipeline::worktree_filter
        PASS [   0.054s] gix-merge::merge blob::pipeline::above_large_file_threshold
        PASS [   0.052s] gix-merge::merge blob::platform::merge::builtin_with_conflict
        PASS [   0.054s] gix-merge::merge blob::platform::merge::builtin_text_uses_binary_if_needed
        PASS [   0.058s] gix-merge::merge blob::platform::merge::one_buffer_too_large
        PASS [   0.070s] gix-merge::merge blob::platform::merge::same_binaries_do_not_count_as_conflicted
        PASS [   0.067s] gix-merge::merge blob::platform::prepare_merge::ancestor_and_current_and_other_do_not_exist
        PASS [   0.067s] gix-merge::merge blob::platform::prepare_merge::driver_selection
        PASS [   0.059s] gix-merge::merge blob::platform::set_resource::invalid_resource_types
        PASS [   0.200s] gix-merge::merge blob::platform::merge::missing_buffers_are_empty_buffers
        FAIL [   0.245s] gix-merge::merge blob::platform::merge::with_external
──── STDOUT:             gix-merge::merge blob::platform::merge::with_external

running 1 test
test blob::platform::merge::with_external ... FAILED

failures:

failures:
    blob::platform::merge::with_external

test result: FAILED. 0 passed; 1 failed; 0 ignored; 0 measured; 19 filtered out; finished in 0.18s

──── STDERR:             gix-merge::merge blob::platform::merge::with_external
thread 'blob::platform::merge::with_external' panicked at gix-merge\tests\merge\blob\platform.rs:297:13:
b
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

        PASS [   0.330s] gix-merge::merge blob::builtin_driver::text::run_baseline
        PASS [   0.998s] gix-merge::merge tree::run_baseline
────────────
     Summary [   1.036s] 20 tests run: 19 passed, 1 failed, 0 skipped
        FAIL [   0.245s] gix-merge::merge blob::platform::merge::with_external
error: test run failed
```

The failing assertion is the first of the assertions in this fragment of the failing test:

https://github.com/GitoxideLabs/gitoxide/blob/2efce72ada61149ef3823f4e15fdbc75157745ec/gix-merge/tests/merge/blob/platform.rs#L291-L314

The first three of the cleaned driver lines are supposed to contain `.tmp`, but the first one is just `b`, which is expected to apper later. This somehow varies based on how the tests are run, and there are at least three environments for figuring out what's going on: CI, local in Git Bash, and local in PowerShell. It seems like they shouldn't differ, but they do.

Locally, it is possible to inspect things in a debugger, but this is harder on CI because while [action-tmate](https://github.com/mxschmitt/action-tmate) can be used to get a reverse shell, on Windows runners the reverse shell's environment differs from the noninteractive environment in which the tests run. Fortunately, I can break the tests in a way that makes them always fail by panicking and printing all the lines:

```diff
diff --git a/gix-merge/tests/merge/blob/platform.rs b/gix-merge/tests/merge/blob/platform.rs
index 6230377b8..62882cc69 100644
--- a/gix-merge/tests/merge/blob/platform.rs
+++ b/gix-merge/tests/merge/blob/platform.rs
@@ -293,6 +293,11 @@ theirs
         let res = platform_ref.merge(&mut buf, default_labels(), &Default::default())?;
         assert_eq!(res, (Pick::Buffer, Resolution::Complete), "merge drivers always merge ");
         let mut lines = cleaned_driver_lines(&buf)?;
+        panic!(
+            "original: {:#?}\ncleaned: {:#?}",
+            buf.lines().map(|s| s.as_bstr()).collect::<Vec<_>>(),
+            lines.collect::<Vec<_>>()
+        );
         for tmp_file in lines.by_ref().take(3) {
             assert!(tmp_file.contains_str(&b".tmp"[..]), "{tmp_file}");
         }
```

To be clear, while all of the following are from messages printed in failing tests, the tests fail *because* of that change. All of them would otherwise pass, except the local run, in Windows, from PowerShell, shown last. Also, I do not know why this is happening. I am hoping you may have some insight into it.

#### Based off the main branch, on CI, in Ubuntu (would pass)

In the [test-fast (ubuntu-latest)](https://github.com/EliahKagan/gitoxide/actions/runs/13406804999/job/37448117443#step:8:1508) run, for comparison:

```text
original: [
    "ours/home/runner/work/gitoxide/gitoxide/gix-merge/.tmpmK4t1a",
    "/home/runner/work/gitoxide/gitoxide/gix-merge/.tmpNFsS5l",
    "/home/runner/work/gitoxide/gitoxide/gix-merge/.tmp8rvAjw",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
cleaned: [
    "ours/home/runner/work/gitoxide/gitoxide/gix-merge/.tmpmK4t1a",
    "/.tmpNFsS5l",
    "/.tmp8rvAjw",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
```

#### Based off the main branch, on CI, in Windows (would pass)

In the [test-fast (windows-latest)](https://github.com/EliahKagan/gitoxide/actions/runs/13406804999/job/37448116375#step:8:1694) run:

```text
original: [
    "oursD:agitoxidegitoxidegix-merge.tmpDYZgPn",
    "D:agitoxidegitoxidegix-merge.tmpOKiCuJ",
    "D:agitoxidegitoxidegix-merge.tmpOahQgz",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
cleaned: [
    "oursD:agitoxidegitoxidegix-merge.tmpDYZgPn",
    "D:agitoxidegitoxidegix-merge.tmpOKiCuJ",
    "D:agitoxidegitoxidegix-merge.tmpOahQgz",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
```

Note how backslashes are being treated as escape characters and removed, which seems like it is not intended, but this does not in and of itself cause a problem with this test: it happens even on the main branch, and even without the changes from this PR.

#### Based off the main branch, locally, in Windows, from Git Bash (would pass)

Via [`cargo nextest run -p gix-merge --no-fail-fast`](https://gist.github.com/EliahKagan/5ea1be801f9770d48826dd1a4df3a3c7):

```text
original: [
    "oursC:Userseksourcereposgitoxidegix-merge.tmpORUoAj",
    "C:Userseksourcereposgitoxidegix-merge.tmpuUGgiZ",
    "C:Userseksourcereposgitoxidegix-merge.tmpsYzC12",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
cleaned: [
    "oursC:Userseksourcereposgitoxidegix-merge.tmpORUoAj",
    "C:Userseksourcereposgitoxidegix-merge.tmpuUGgiZ",
    "C:Userseksourcereposgitoxidegix-merge.tmpsYzC12",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
```

#### Based off the main branch, locally, in Windows, from PowerShell (would pass)

Via [`cargo nextest run -p gix-merge --no-fail-fast`](https://gist.github.com/EliahKagan/a6b8584526b5153c7bedf588e2e0fdfd):

```text
original: [
    "oursC:Userseksourcereposgitoxidegix-merge.tmpPNy4Pa",
    "C:Userseksourcereposgitoxidegix-merge.tmp5Fkmzy",
    "C:Userseksourcereposgitoxidegix-merge.tmp9aFhwO",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
cleaned: [
    "oursC:Userseksourcereposgitoxidegix-merge.tmpPNy4Pa",
    "C:Userseksourcereposgitoxidegix-merge.tmp5Fkmzy",
    "C:Userseksourcereposgitoxidegix-merge.tmp9aFhwO",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
```

#### Based off this feature branch, on CI, in Ubuntu (would pass)

In the [test-fast (ubuntu-latest)](https://github.com/EliahKagan/gitoxide/actions/runs/13406769468/job/37448018317#step:8:1508) run, for comparison:

```text
original: [
    "ours/home/runner/work/gitoxide/gitoxide/gix-merge/.tmpvKkoxy",
    "/home/runner/work/gitoxide/gitoxide/gix-merge/.tmpr402lE",
    "/home/runner/work/gitoxide/gitoxide/gix-merge/.tmp8bJjIe",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
cleaned: [
    "ours/home/runner/work/gitoxide/gitoxide/gix-merge/.tmpvKkoxy",
    "/.tmpr402lE",
    "/.tmp8bJjIe",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
```

That is the same as based off the main branch, which is very much as expected, because the changes on this feature branch do not produce a different value for `gix_path::env::shell()` on systems other than Windows.

#### Based off this feature branch, on CI, on Windows (would pass)

In the [test-fast (windows-latest)](https://github.com/EliahKagan/gitoxide/actions/runs/13406769468/job/37448017722#step:8:1559) run:

```text
original: [
    "oursD:agitoxidegitoxidegix-merge.tmpyuHJtT",
    "D:agitoxidegitoxidegix-merge.tmpC2s55H",
    "D:agitoxidegitoxidegix-merge.tmp06rraO",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
cleaned: [
    "oursD:agitoxidegitoxidegix-merge.tmpyuHJtT",
    "D:agitoxidegitoxidegix-merge.tmpC2s55H",
    "D:agitoxidegitoxidegix-merge.tmp06rraO",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
```

This is similar to the main branch CI and local Git Bash runs shown above.

#### Based off this feature branch, locally, in Windows, from Git Bash (would pass)

Via [`cargo nextest run -p gix-merge --no-fail-fast`](https://gist.github.com/EliahKagan/3dbe97b9e291ee5ec66538379fb2503b):

```text
original: [
    "oursC:Userseksourcereposgitoxidegix-merge.tmpUfxnE8",
    "C:Userseksourcereposgitoxidegix-merge.tmpiUdSfX",
    "C:Userseksourcereposgitoxidegix-merge.tmpmkVmRq",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
cleaned: [
    "oursC:Userseksourcereposgitoxidegix-merge.tmpUfxnE8",
    "C:Userseksourcereposgitoxidegix-merge.tmpiUdSfX",
    "C:Userseksourcereposgitoxidegix-merge.tmpmkVmRq",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
    "b",
    "theirs",
]
```

#### Based off this feature branch, locally, in Windows, from PowerShell (would fail)

Via [`cargo nextest run -p gix-merge --no-fail-fast`](https://gist.github.com/EliahKagan/1597402f5e78a0056b76f903a477ce35):

```text
original: [
    "b",
    "theirserseksourcereposgitoxidegix-merge.tmpTDKDff",
    "C:Userseksourcereposgitoxidegix-merge.tmp5WJaSb",
    "C:Userseksourcereposgitoxidegix-merge.tmpyQDnZt",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
]
cleaned: [
    "b",
    "theirserseksourcereposgitoxidegix-merge.tmpTDKDff",
    "C:Userseksourcereposgitoxidegix-merge.tmp5WJaSb",
    "C:Userseksourcereposgitoxidegix-merge.tmpyQDnZt",
    "7",
    "b",
    "ancestor label",
    "current label",
    "other label",
    "%F",
]
```

Note how `b` comes first, and the three lines with paths to files named containing `.tmp` follow it. This causes this assertion, which is run for the first three lines, to fail in the first iteration (i.e. for the first line):

https://github.com/GitoxideLabs/gitoxide/blob/2efce72ada61149ef3823f4e15fdbc75157745ec/gix-merge/tests/merge/blob/platform.rs#L297

But I do not understand why. I'm not very familiar with the details of `gix-merge`. I'm also not entirely clear on which kinds of shell transformations (from parsing `%` items into fields, from shell expansions on unquoted parameter expansion, and from the implicit effect of joining on spaces when passing multiple arguments arising from an unquoted expansion to `echo`) are intended to occur in the test command, and with what effects.

…