@Byron 2025-02-24
Thanks you very much for tackling this!

I didn't look at the code of this PR but want to share some thoughts on the failure.

Here is the program that is run:

```bash
for arg in  %O %A %B %L %P %S %X %Y %F; do echo $arg >> \"%A\"; done; cat \"%O\" \"%B\" >> \"%A\""
```

The failing output has 10 lines, whereas a valid output has 11 lines. Probably that is related. My feeling is that one of the arguments in the for loop substitutions  is empty. It would disappear then, and everything is offset.
If these were quoted that shouldn't happen and is probably a good idea just to make the test-script more robust.

From there it should become more obvious which parameter comes out empty, and from there it should become clear what the problem really is.

I hope that helps.