@Byron 2025-02-25
`theirs` is empty, even though it should be a be something that the caller knows. 

> I am very unclear on what the exact intended behavior of this merge driver test command is.

It attempts to make the arguments that the merge-driver was called with observable. I think there is no quoting because each of the arguments is known not to have spaces in it. However, the lack of quoting also makes empty arguments unobservable, a definitive shortcoming of the current implementation.

> The broader issue for me examining this specific test is that I am familiar neither with the details of `gix-merge` and its tests nor (even though this is general Git knowledge) with the format expected for custom merge drivers.

[Here is the docs](https://git-scm.com/docs/gitattributes#_defining_a_custom_merge_driver), but the short version is that it's a script into which various parameters can be substituted into - this substitution is performed by the the caller of the merge-driver. That caller does a bare string substitution, without any care for quotes or whitespace in general.

> I'd like to avoid become too embroiled in the details of the gix-merge tests.

I definitely wouldn't want you to suffer unnecessarily. The key for me to solve this certainly is to reproduce the error. For now, it doesn't make much sense to me either, unless… it's a substitution engine, what if one of the arguments that it substitutes also contains substitutable characters? Probably that's not what's happening here though, as `%` should be uncommon enough.

This is a patch that should tell exactly what's executed, and from there it would become more obvious what's really happening.

```patch
diff --git a/gix-merge/src/blob/platform/merge.rs b/gix-merge/src/blob/platform/merge.rs
index 058a30128..324742a00 100644
--- a/gix-merge/src/blob/platform/merge.rs
+++ b/gix-merge/src/blob/platform/merge.rs
@@ -211,7 +211,7 @@ pub(super) mod inner {
                     cmd.extend_from_slice(token);
                 }
 
-                Ok(merge::Command {
+                Ok(dbg!(merge::Command {
                     cmd: gix_command::prepare(gix_path::from_bstring(cmd))
                         .with_context(context)
                         .command_may_be_shell_script()
@@ -223,7 +223,7 @@ pub(super) mod inner {
                     current_path: ours_path,
                     ancestor: base_tmp,
                     other: theirs_tmp,
-                })
+                }))
             }
 
             /// Return the configured driver program for use with [`Self::prepare_external_driver()`], or `Err`
```

I hope that helps.